refactor: runnerpost 테스트 패키징 변경
